Update test timeouts
The upgrate tests time out while waiting for the cilium image
to be downloaded.

Note that we have a large amount of parallel tests (~100) that
do not share the same env and cannot benefit from the image
proxy, leading to a large number of image downloads.

For now, we'll increase the test timeouts.