Create missing snap tracks if necessary
This commit ensures the creation of snap tracks for a release if they do not already exist.
It interacts directly with the charmhub.io API (not snapstore.io) to create any missing tracks.
Only new tracks that conform to the naming guard-rails (1.XX[-classic|strict|moonray]) will be automatically created.